Refactor: move module
