+++ pkg/registry/auth/credentials.go
@@ -1,4 +1,4 @@
-package registry_auth
+package auth
 
 import (
 	"fmt"
