+++ pkg/registry/auth/dockerfiles.go
@@ -1,4 +1,4 @@
-package registry_auth
+package auth
 
 import (
 	"encoding/base64"
